+++ docs/pages/changelog.md
@@ -10,6 +10,7 @@ Summary of changes.
     => eg now easier to to `ontospy serialize foaf.rdf json-ld > ~/foaf.json`
 -   embedded ontodocs plugin within main source code
     -   requires install with [HTML] flag
+-   'utils' and 'quickdocs' commands added to CLI
 
 ## November 12, 2018: v1.9.6